# api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/article/updatetags": {
      "put": {
        "tags": [
          "article-controller"
        ],
        "operationId": "updateTags",
        "parameters": [
          {
            "name": "articleId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "uniqueItems": true,
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TagDTO"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleDTO"
                }
              }
            }
          }
        }
      }
    },
    "/article/update": {
      "put": {
        "tags": [
          "article-controller"
        ],
        "operationId": "updateArticle",
        "parameters": [
          {
            "name": "articleId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/bookmark/add": {
      "post": {
        "tags": [
          "bookmarked-controller"
        ],
        "operationId": "addBookmark",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "articleId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/article/save": {
      "post": {
        "tags": [
          "article-controller"
        ],
        "operationId": "saveArticle",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleDTO"
                }
              }
            }
          }
        }
      }
    },
    "/user/{userId}": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/comment/article/{articleId}": {
      "get": {
        "tags": [
          "comment-controller"
        ],
        "operationId": "getCommentsByArticleId",
        "parameters": [
          {
            "name": "articleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/bookmark/user/{userId}": {
      "get": {
        "tags": [
          "bookmarked-controller"
        ],
        "operationId": "getBookmarkedArticlesByUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/article/{articleId}": {
      "get": {
        "tags": [
          "article-controller"
        ],
        "operationId": "getArticleById",
        "parameters": [
          {
            "name": "articleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleDTO"
                }
              }
            }
          }
        }
      }
    },
    "/article/title": {
      "get": {
        "tags": [
          "article-controller"
        ],
        "operationId": "getArticlesByTitle",
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/article/tag/{tag}": {
      "get": {
        "tags": [
          "article-controller"
        ],
        "operationId": "getArticlesByTag",
        "parameters": [
          {
            "name": "tag",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/article/published": {
      "get": {
        "tags": [
          "article-controller"
        ],
        "operationId": "getArticlesByPublishedDate",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/article/comments/{articleId}": {
      "get": {
        "tags": [
          "article-controller"
        ],
        "operationId": "getCommentsForArticle",
        "parameters": [
          {
            "name": "articleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/article/author": {
      "get": {
        "tags": [
          "article-controller"
        ],
        "operationId": "getArticlesByAuthor",
        "parameters": [
          {
            "name": "authorId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/article/author/premium": {
      "get": {
        "tags": [
          "article-controller"
        ],
        "operationId": "getByAuthorAndPremium",
        "parameters": [
          {
            "name": "authorId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "premiumRequired",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/bookmark": {
      "delete": {
        "tags": [
          "bookmarked-controller"
        ],
        "operationId": "deleteBookmark",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "articleId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/article/delete": {
      "delete": {
        "tags": [
          "article-controller"
        ],
        "operationId": "deleteArticle",
        "parameters": [
          {
            "name": "articleId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ArticleDTO": {
        "type": "object",
        "properties": {
          "articleId": {
            "type": "integer",
            "format": "int32"
          },
          "authorId": {
            "type": "integer",
            "format": "int32"
          },
          "publishedDate": {
            "type": "string",
            "format": "date-time"
          },
          "premiumRequired": {
            "type": "boolean"
          },
          "content": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "tags": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagDTO"
            }
          },
          "likes": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LikesDTO"
            }
          },
          "bookmarks": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookmarkedDTO"
            }
          },
          "comments": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentDTO"
            }
          }
        }
      },
      "BookmarkedDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userDTO": {
            "$ref": "#/components/schemas/UserDTO"
          },
          "articleDTO": {
            "$ref": "#/components/schemas/ArticleDTO"
          }
        }
      },
      "CommentDTO": {
        "type": "object",
        "properties": {
          "commentId": {
            "type": "integer",
            "format": "int64"
          },
          "content": {
            "type": "string"
          },
          "userDTO": {
            "$ref": "#/components/schemas/UserDTO"
          },
          "articleDTO": {
            "$ref": "#/components/schemas/ArticleDTO"
          },
          "parentCommentId": {
            "type": "integer",
            "format": "int64"
          },
          "createdOn": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FollowRelationsDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "whoIsFollowing": {
            "$ref": "#/components/schemas/UserDTO"
          },
          "whoIsBeingFollowed": {
            "$ref": "#/components/schemas/UserDTO"
          }
        }
      },
      "LikesDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userDTO": {
            "$ref": "#/components/schemas/UserDTO"
          },
          "articleDTO": {
            "$ref": "#/components/schemas/ArticleDTO"
          }
        }
      },
      "TagDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "tag": {
            "type": "string"
          },
          "articleDTO": {
            "$ref": "#/components/schemas/ArticleDTO"
          }
        }
      },
      "UserDTO": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "emailId": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "likes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LikesDTO"
            }
          },
          "bookmarks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BookmarkedDTO"
            }
          },
          "followers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FollowRelationsDTO"
            }
          },
          "following": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FollowRelationsDTO"
            }
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentDTO"
            }
          },
          "subscribed": {
            "type": "boolean"
          },
          "bio": {
            "type": "string"
          }
        }
      }
    }
  }
}